Refactor to use graphql-binding and to enable static bindings
