+++ package.json
@@ -1,21 +1,24 @@
 {
   "name": "graphcool-binding",
   "version": "0.3.1",
-  "main": "dist/src/index.js",
-  "typings": "./dist/src/index.d.ts",
+  "main": "dist/index.js",
+  "typings": "./dist/index.d.ts",
   "files": [
     "dist"
   ],
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/graphcool/graphcool-orm.git"
+    "url": "git+https://github.com/graphcool/graphcool-binding.git"
   },
   "author": "Johannes Schickling <johannes@graph.cool>",
+  "contributors": [
+    "Kim Brandwijk <kim.brandwijk@gmail.com>"
+  ],
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/graphcool/graphcool-orm/issues"
+    "url": "https://github.com/graphcool/graphcool-binding/issues"
   },
-  "homepage": "https://github.com/graphcool/graphcool-orm",
+  "homepage": "https://github.com/graphcool/graphcool-binding",
   "scripts": {
     "prepublish": "npm run build",
     "build": "rm -rf dist && tsc -d",
@@ -30,7 +33,6 @@
     "graphql": "^0.12.3",
     "graphql-binding": "^0.1.0",
     "graphql-import": "^0.1.8",
-    "graphql-request": "^1.4.0",
     "graphql-schema-cache": "^0.3.2",
     "graphql-tools": "^2.14.1",
     "jsonwebtoken": "^8.1.0"
@@ -39,8 +41,8 @@
     "@types/graphql": "^0.11.7",
     "@types/node": "^8.5.2",
     "ava": "^0.24.0",
-    "tslint": "^5.5.0",
+    "tslint": "^5.8.0",
     "tslint-config-standard": "^7.0.0",
-    "typescript": "^2.4.2"
+    "typescript": "^2.6.2"
   }
 }
